+++ lib/controller/abstract.js
@@ -1,6 +1,6 @@
 var path = require('path');
 var util = require('util');
-var support = require(path.join(__dirname, 'support'));
+var support = require('../support');
 var _ = support.Utilities;
 var reflectAccessors = support.ReflectAccessors;
 var delegates = require('delegates');
@@ -12,36 +12,40 @@ var delegates = require('delegates');
  *
  * @param {Context} ctx Koa.Context instance from the request
  */
-function AbstractController(ctx) {
+function AbstractController() {
   this.locals = {};
 }
 
 /*
  * Prototype
  */
+_.extend(AbstractController.prototype,
+  require('./filters'),
+  require('./rendering')
+);
 _.extend(AbstractController.prototype, {
-
   /**
-   * simply write the argument[0] to the response body
-   * all template / advanced rendering is done in kona-rendering
-   *
-   * @param {Mixed} data          data to set to ctx.body
-   * @param {String} options       options including type: 'content-type' to
-   *                               to set to ctx.type
+   * set a variable (or variables) to the locals that will be provided to the
+   * view.
+   * @param {Object|String} objectOrKey an object of variable name => value map
+   *                                    or a string that will be the variable name
+   *                                    in the view.
+   * @param {*} value       the value to store as the named variable [optional]
    */
-  render: function* (data, options) {
+  set: function(objectOrKey, value) {
 
-    options || (options = {});
+    if (_.isPlainObject(objectOrKey)) {
 
-    if (options.type) {
-      this.type = options.type;
+      _.assign(this.locals, objectOrKey);
+
+    } else if (_.isString(objectOrKey) && !_.isUndefined(value)) {
+
+      this.locals[objectOrKey] = value;
+
     }
-
-    this.body = data;
   }
 
 });
-
 
 /**
  * an extend 'class' method to create child / sub 'classes' of this conroller
@@ -51,9 +55,9 @@ _.extend(AbstractController.prototype, {
  * example:
  *
  * ```
- *   var ApplicationController = kona.Controller.Base.extend({
+ *   var ApplicationController = kona.Controller.Abstract.extend({
  *     constructor: function() {
- *       kona.Controller.Base.apply(this, arguments);
+ *       kona.Controller.Abstract.apply(this, arguments);
  *       this.beforeFilter('authenticate');
  *     }
  *   })
@@ -75,6 +79,12 @@ AbstractController.extend = require('class-extend').extend;
  * @return {Function}             the constructor that inherits from this.
  */
 AbstractController.rebase = _.partialRight(util.inherits, AbstractController);
+
+
+/**
+ * Statics
+ */
+_.extend(AbstractController, require('./renderers'));
 
 
 module.exports = AbstractController;
