+++ lib/kona.js
@@ -7,7 +7,7 @@ var repl = require('repl');
 var util = require('util');
 var inherits = util.inherits;
 var format = util.format;
-var support = require(join(__dirname, 'support'));
+var support = require('./support');
 var LivePath = support.LivePath;
 var chalk = require('chalk');
 var _ = require('lodash');
@@ -42,10 +42,12 @@ inherits(Kona, Koa);
 /**
  * extend Kona.prototype
  */
-_.extend(Kona.prototype, require(join(__dirname, 'kona', 'logger')));
-_.extend(Kona.prototype, require(join(__dirname, 'kona', 'watch')));
-_.extend(Kona.prototype, require(join(__dirname, 'kona', 'mixins')));
-_.extend(Kona.prototype, require(join(__dirname, 'kona', 'hooks')));
+_.extend(Kona.prototype,
+  require('./kona/logger'),
+  require('./kona/watch'),
+  require('./kona/mixins'),
+  require('./kona/hooks')
+);
 _.extend(Kona.prototype, {
 
   /**
@@ -225,9 +227,24 @@ _.extend(Kona.prototype, {
    * within the application
    */
   mountBaseModules: function () {
-    this.Controller = {
-      Base: require(this._root.join('lib', 'controller', 'base'))
-    };
+
+    Object.defineProperty(this, 'Controller', {
+      writable: false,
+      value: {
+        Request: require('./controller/request'),
+        Abstract: require('./controller/abstract')
+      }
+    });
+
+    Object.defineProperty(this.Controller, 'Base', {
+      get: function() {
+        console.log([
+          'kona.Controller.Base has been deprecated. Please use',
+          'kona.Controller.Request for request controller extension.'
+        ].join(' '));
+        return this.Request;
+      }
+    })
   },
 
 
@@ -277,7 +294,7 @@ _.extend(Kona.prototype, {
       throw new Error('Cannot call #listen before Kona has been intialized');
     }
 
-    bean = require(join(__dirname, 'utilities', 'bean'));
+    bean = require('./utilities/bean');
 
     if (this.env === 'test') {
       delete this.port;
