+++ lib/kona/mixins.js
@@ -66,7 +66,7 @@ module.exports = {
       var modulePath = path.join(loadPath, folder),
           mixin;
 
-      if (/^kona\-.*/.test(folder) && fs.statSync(folder).isDirectory()) {
+      if (/^kona\-.*/.test(folder) && fs.statSync(modulePath).isDirectory()) {
 
         this.requireMixin(modulePath, folder);
