+++ lib/middleware/dispatcher.js
@@ -62,10 +62,10 @@ module.exports = function(app) {
           )));
 
         // controller module does not inherit from kona controller/base
-        } else if (!this.controller instanceof kona.Controller.Base) {
+        } else if (!this.controller instanceof kona.Controller.Request) {
 
           return this.throw(500, new Error(format(
-            "%s does not inherit from Kona.Controller.Base",
+            "%s does not inherit from Kona.Controller.Request",
             this.controller
           )));
