+++ test/lib/controller/request.test.js
@@ -3,8 +3,8 @@ var chai = require('chai');
 var expect = chai.expect;
 var sinon = require('sinon');
 chai.use(require('sinon-chai'));
-var Kona = require(path.join(__dirname.replace(/test\//, ''), 'kona'));
-var AbstractController = require(__filename.replace(/test(\/|\.)?/g, ''));
+var Kona = require(path.join(__dirname.replace(/test\//, ''), '..', 'kona'));
+var RequestController = require(__filename.replace(/test(\/|\.)?/g, ''));
 var co = require('co');
 var Writable = require('stream').Writable;
 
@@ -15,81 +15,13 @@ describe('controller/base', function() {
     app = new Kona();
   });
 
-  describe('#set', function() {
-
-    var controller, ctx;
-
-    beforeEach(function() {
-      ctx = getCtx(app),
-      controller = new AbstractController(ctx);
-    });
-
-    describe('given a key-value pair', function() {
-
-      it('sets the value at the key to ctx.locals', function() {
-
-        var key = 'test',
-            value = {one: 'two'};
-        controller.set(key, value);
-
-        expect(controller.locals).to.include.key(key);
-        expect(controller.locals[key]).to.eql(value);
-        expect(ctx.locals[key]).to.eql(value);
-
-      });
-
-      it('sets an object map of key values paris to ctx.locals', function() {
-
-        var object = {
-          one: 'two',
-          three: 'four'
-        };
-        controller.set(object);
-
-        expect(controller.locals).to.include.keys(Object.keys(object));
-        Object.keys(object).forEach(function(key) {
-          expect(ctx.locals[key]).to.eql(object[key]);
-          expect(controller.locals[key]).to.eql(object[key]);
-        });
-
-      });
-
-      it('assigns object key value pairs to locals without overwriting existing', function() {
-
-        var object = {
-          one: 'two',
-          three: 'four'
-        };
-        controller.locals['existing'] = true;
-
-        controller.set(object);
-        expect(controller.locals.existing).to.be.true;
-        expect(ctx.locals.existing).to.be.true;
-
-        Object.keys(object).forEach(function(key) {
-          expect(ctx.locals[key]).to.eql(object[key]);
-          expect(controller.locals[key]).to.eql(object[key]);
-        });
-
-      });
-
-    })
-
-  });
-
   describe('#respondTo', function() {
 
     var controller, ctx;
 
     beforeEach(function() {
-      ctx = {locals: {},
-        request: {
-          accepts: function() {}
-        },
-        router: {},
-        throw: function() {}
-      };
-      controller = new AbstractController(ctx);
+      ctx = getCtx(app);
+      controller = new RequestController(ctx);
     });
 
     describe('when no type: *responder object is given', function() {
@@ -188,7 +120,7 @@ describe('controller/base', function() {
 
     beforeEach(function() {
       ctx = getCtx(),
-      controller = new AbstractController(ctx);
+      controller = new RequestController(ctx);
     });
 
     it('adds the respond content-type to an private array', function() {
@@ -211,7 +143,7 @@ describe('controller/base', function() {
 
     beforeEach(function() {
       ctx = getCtx(app);
-      controller = new AbstractController(ctx, kona);
+      controller = new RequestController(ctx, kona);
     });
 
     xit('stores the render call arguments for the responder', function() {
@@ -241,89 +173,20 @@ describe('controller/base', function() {
 
   });
 
-  describe('#{before|after}Filter', function() {
-
-    var controller, ctx;
-
-    beforeEach(function() {
-      ctx = getCtx(app),
-      controller = new AbstractController(ctx);
-    });
-
-    it('registers {before|after}Filter functions', function() {
-
-      controller.beforeFilter('one');
-      expect(controller._beforeFilters.length).to.eq(1);
-      expect(controller._beforeFilters[0]).to.eq('one');
-      controller.beforeFilter('two', 'three');
-      expect(controller._beforeFilters.length).to.eq(3);
-      expect(controller._beforeFilters).to.eql(['one', 'two', 'three']);
-
-    });
-
-  });
-
-  describe('#call{Before|After}Filters', function() {
-
-    var controller, ctx;
-
-    beforeEach(function() {
-      ctx = getCtx(),
-      controller = new (AbstractController.extend({
-        before: function*() {},
-        after: function*() {}
-      }))(ctx);
-    });
-
-    it('calls all before filters', function() {
-
-      var spy = sinon.spy(Object.getPrototypeOf(controller), 'before');
-      controller.beforeFilter('before');
-      var gen = controller.callBeforeFilters();
-      gen.next();
-      expect(spy).to.be.calledOnce;
-
-    });
-
-    it('calls all after filters', function() {
-
-      var spy = sinon.spy(Object.getPrototypeOf(controller), 'after');
-      controller.afterFilter('after');
-      var gen = controller.callAfterFilters();
-      gen.next();
-      expect(spy).to.be.calledOnce;
-    });
-
-
-    it('does nothing when there are no filters', function() {
-
-      var gen, result;
-      expect(controller._beforeFilters).to.be.empty;
-      expect(controller._afterFilters).to.be.empty;
-      gen = controller.callAfterFilters();
-      result = gen.next();
-      expect(result.value).to.be.undefined;
-      gen = controller.callBeforeFilters();
-      result = gen.next();
-      expect(result.value).to.be.undefined;
-
-    });
-
-  });
 
   describe('::addRenderer', function() {
     it('explodes when type is not given and render is not a function', function() {
 
       expect(function() {
-        AbstractController.addRenderer(function() {});
+        RequestController.addRenderer(function() {});
       }).to.throw(Error);
 
       expect(function() {
-        AbstractController.addRenderer('josn', function() {});
+        RequestController.addRenderer('josn', function() {});
       }).to.throw(Error);
 
       expect(function() {
-        AbstractController.addRenderer('josn', function*() {});
+        RequestController.addRenderer('josn', function*() {});
       }).to.not.throw(Error);
 
     });
@@ -334,9 +197,9 @@ describe('controller/base', function() {
       var renderer = function*() {
         return yield '|_|_|1|_|3|_|_|1|_|';
       };
-      AbstractController.addRenderer('xlsx', renderer);
+      RequestController.addRenderer('xlsx', renderer);
 
-      expect(AbstractController.getRenderer('xlsx')).to.eq(renderer);
+      expect(RequestController.getRenderer('xlsx')).to.eq(renderer);
     });
   });
 
@@ -353,7 +216,7 @@ describe('controller/base', function() {
     Object.keys(map).forEach(function(type) {
       describe(type, function() {
         it('returns their input when handled by koa', function() {
-          var renderer = AbstractController.getRenderer(type),
+          var renderer = RequestController.getRenderer(type),
               data = map[type];
 
           co.wrap(renderer)(data).then(function(returned) {
@@ -370,7 +233,7 @@ describe('controller/base', function() {
 
       var error,
           ctx = getCtx(app),
-          ctrl = new AbstractController(ctx),
+          ctrl = new RequestController(ctx),
           spy = sinon.spy(Object.getPrototypeOf(ctrl), 'respondTo');
           acceptSpy = sinon.stub(ctrl.request, 'accepts').returns();
